fix Dev.json + param.json error
This is a hack based on the fact that param.json now includes a "slotName" parameter to enable the TiP scenario. We make the template file Dev.json work with the parameter file (instead of the other way around like we normally do) for simplicity sake of the tutorials, even though it doesn't actually use the "slotName" parameter.